Read the JRTE Corpus

Download and read the Japanese Realistic Textual Entailment Corpus. The result of this function is memoised with memoise::memoise() internally.

Usage

read_jrte(
  url = "https://github.com/megagonlabs/jrte-corpus/archive/refs/heads/master.zip",
  exdir = tempdir(),
  keep = jrte_rte_files(),
  keep_rhr = FALSE,
  keep_pn = FALSE
)

Arguments

url

String. If left with NULL, the function will skip downloading the file.

exdir

String. Path to tempolarily unzip text files.

keep

List. File names to parse and keep in returned value.

keep_rhr

Logical. If supplied TRUE, keeps rhr.tsv.

keep_pn

Logical. If supplied TRUE, keeps pn.tsv.

Value

A list of tibbles.

See also